requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10 
sqlparse==0.5.3
typing_extensions==4.12.2
django-allauth==0.63.6
google-auth==2.35.0
django-cors-headers==4.3.1
requests==2.32.3
PyJWT==2.9.0
cryptography==43.0.3